Refactor out a hook for creating new graph via Graph API
